// src/types/article.ts
export interface Article {
  id: string;
  slug: string;
  title: string;
  description: string;
  category: string;
  content: string;
  lastUpdated: string;
}

export interface ArticleMetadata {
  id: string;
  slug: string;
  title: string;
  description: string;
  category: string;
  lastUpdated: string;
} 